services/data_service/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
sqlalchemy>=2.0.9
psycopg2-binary>=2.9.6
yfinance>=0.2.18
requests>=2.28.2
python-dotenv>=1.0.0
python-jose>=3.3.0
passlib>=1.7.4
